# Conflicts:
#	src/main/java/za/ac/cput/repository/IBookingRepository.java
#	src/main/java/za/ac/cput/repository/ICustomerRepository.java
#	src/main/java/za/ac/cput/repository/IEmployeeRepository.java

+++ src/main/java/za/ac/cput/factory/EmployeeFactory.java
@@ -13,7 +13,7 @@ public class EmployeeFactory {
         String employeeCode = Helper.generateID();
 
         return new Employee.Builder()
-                .setId(id)
+                .setId((int) id)
                 .setFirstName(firstName)
                 .setLastName(lastName)
                 .setEmployeeCode(employeeCode)
